flower mode

+++ burningship.c
@@ -14,14 +14,14 @@
 
 void	ft_iterative_ship(t_mlx *list)
 {
-	float	tmp;
+	double		tmp;
 
 	while (ITR < ITRMAX && ((FLOW == 0 && ZR * ZR + ZI * ZI < 100)
 		|| (FLOW == 1 && ZR * (ZI + ZI) * (ZI + ZI) * ZR < 20)
 			|| (FLOW == 2 && ZI * ZR + ZR * ZI < 40)))
 	{
-		tmp = fabsf(ZR * ZR - ZI * ZI + CR);
-		ZI = fabsf(2 * ZR * ZI + CI);
+		tmp = fabs(ZR * ZR - ZI * ZI + CR);
+		ZI = fabs(2 * ZR * ZI + CI);
 		ZR = tmp;
 		ITR += 1;
 	}
